# perf: decrease run time by ~40% by avoiding allocations in hot loop