While bench marking `2fst` I notice that a lot of time is spend allocating and de-allocating `Vec`s in `read_value_changes`. I avoided those allocations getting slices from `mu_slice`, instead of reading it to a new `Vec`, and by adding a reusable buffer for `multi_bit_digital_signal_to_chars`.

I added a new example for making the speed measurements. The only `.fst` file with a reasonable size I have is the one I got from running `vcd2fst` on `trace.vcd`, not sure how representative it is.

[trace.zip](https://github.com/user-attachments/files/18231536/trace.zip)

```
> hyperfine './read-6ec01e3 trace.fst' './read-f162357 trace.fst'
Benchmark 1: ./read-6ec01e3 trace.fst
  Time (mean ± σ):      1.191 s ±  0.029 s    [User: 1.127 s, System: 0.059 s]
  Range (min … max):    1.151 s …  1.235 s    10 runs

Benchmark 2: ./read-f162357 trace.fst
  Time (mean ± σ):     672.5 ms ±  10.4 ms    [User: 611.5 ms, System: 56.9 ms]
  Range (min … max):   653.5 ms … 689.6 ms    10 runs

Summary
  ./read-f162357 trace.fst ran
    1.77 ± 0.05 times faster than ./read-6ec01e3 trace.fst
```